+++ system/modules/photoalbums2/dca/tl_module.php
@@ -408,9 +408,13 @@ $GLOBALS['TL_DCA']['tl_module']['fields']['pa2TimeFilterEnd'] = array(
 $GLOBALS['TL_DCA']['tl_module']['fields']['pa2Teaser'] = array(
     'label'                   => &$GLOBALS['TL_LANG']['tl_module']['pa2Teaser'],
     'exclude'                 => true,
-    'inputType'               => 'TranslationTextArea',
+	'inputType'               => 'textarea',
+	'load_callback'           => array
+	(
+		array('\TranslationFieldsHelper', 'getTranslation')
+	),
     'eval'                    => array('rte' => 'tinyMCE', 'tl_class' => 'long'),
-    'sql'                     => "int(10) unsigned NOT NULL default '0'",
+    'sql'                     => "blob NULL",
 );
 
 /**